// src/lib/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.1 (d3f7cba)";
  };
  public: {
    Tables: {
// ...
      schools: {
        Row: {
          address: string | null;
          city: string;
          code: string;
          created_at: string | null;
          created_by: string | null;
          cycle_id: string;
          email: string;
          id: string;
          image_url: string | null;
          is_technical_education: boolean | null;
          name: string;
          phone: string;
          state_id: number | null;
          status: Database["public"]["Enums"]["school_status_enum"];
          updated_at: string | null;
          updated_by: string | null;
        };
        Insert: {
          address?: string | null;
          city: string;
          code: string;
          created_at?: string | null;
          created_by?: string | null;
          cycle_id: string;
          email: string;
          id?: string;
          image_url?: string | null;
          is_technical_education?: boolean | null;
          name: string;
          phone: string;
          state_id?: number | null;
          status?: Database["public"]["Enums"]["school_status_enum"];
          updated_at?: string | null;
          updated_by?: string | null;
        };
        Update: {
          address?: string | null;
          city?: string;
          code?: string;
          created_at?: string | null;
          created_by?: string | null;
          cycle_id?: string;
          email?: string;
          id?: string;
          image_url?: string | null;
          is_technical_education?: boolean | null;
          name?: string;
          phone?: string;
          state_id?: number | null;
          status?: Database["public"]["Enums"]["school_status_enum"];
          updated_at?: string | null;
          updated_by?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "schools_cycle_id_foreign";
            columns: ["cycle_id"];
            isOneToOne: false;
            referencedRelation: "cycles";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "schools_state_id_foreign";
            columns: ["state_id"];
            isOneToOne: false;
            referencedRelation: "states";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
// ...
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;